policies/latency-routing/scripts/requirements.txt
locust==2.26.0
openai==1.23.3
azure-monitor-opentelemetry==1.3.0
